Move accordion functions from `docassemble.demo` to here
So we don't have an explicit dependency on docassemble. Otherwise,
developing this package on a higher version docassemble server and
then installing it on a different lower version server will break
the lower version server.

Also moved the `license` in the `pyproject.toml` to it's right place
(another docassemble bug).